+++ src/renderer/backends/SoftwareMapFadePass.h
@@ -24,8 +24,8 @@ public:
     SoftwareMapFadePass()  = default;
     ~SoftwareMapFadePass() = default;
 
-    long StepFadeIn(long step) override;
-    long StepFadeOut(long step) override;
+    int32_t StepFadeIn(int32_t step) override;
+    int32_t StepFadeOut(int32_t step) override;
 
     const char* GetName() const override { return "SOFTWARE"; }
 };
